Tweak HUD behaviour and adjust size of checkout zones
End-of-game panel stays up longer for winning results, which now include time remaining at checkout. This currently does not affect score, but may in the future. Additionally, timer readout should now not have any edge cases that will cause issues in formatting a string from the calculations it reads. Made checkout zones in the level a bit smaller so they are not accidentally walked into by the player.